Update PyWake / TopFarm tests with new API
