+++ test/plant/test_system_pywake.py
@@ -1,6 +1,8 @@
-from windIO.utils import plant_examples_data_path
+
+from pathlib import Path
 from windIO.utils.yml_utils import load_yaml
-from windIO.utils.pywake_utils import ymlSystem2PyWake
+import windIO.reference_library.plant as windIO_plant_lib
+from .pywake_utils import ymlSystem2PyWake
 from py_wake.deficit_models.gaussian import IEA37SimpleBastankhahGaussian
 import numpy as np
 import numpy.testing as npt
@@ -13,15 +15,35 @@ try:
     from topfarm.cost_models.py_wake_wrapper import PyWakeAEPCostModelComponent
     from topfarm.constraint_components.boundary import CircleBoundaryConstraint
     from topfarm.plotting import XYPlotComp
-    from windIO.utils.topfarm_utils import ymlSystem2TopFarm
-    from test.plant import examples_data_path
 except ModuleNotFoundError:
     topfarm = None
+
+    def ymlSystem2TopFarm(wind_energy_system_yml, **kwargs):
+        wes = load_yaml(wind_energy_system_yml)
+        wfm, (x, y) = ymlSystem2PyWake(wes, IEA37SimpleBastankhahGaussian)
+        constraints = []
+        boundary = wes['site']['boundaries']
+        for k, v in boundary.items():
+            if k == 'circle':
+                center = [v['center'][xy] for xy in 'xy']
+                radius = v['radius']
+                constraints.append(CircleBoundaryConstraint(center, radius))
+            else:
+                raise NotImplementedError()
+
+        tf_kwargs = {'design_vars': {'x': x, 'y': y},
+                    'cost_comp': PyWakeAEPCostModelComponent(windFarmModel=wfm, n_wt=len(x)),
+                    'constraints': [CircleBoundaryConstraint(center, radius)],
+                    'plot_comp': XYPlotComp()}
+        tf_kwargs.update(kwargs)
+        return TopFarmProblem(**tf_kwargs)
 
     def test_IEA37_case_study1_16wt_PyWake():
         if topfarm is None:
             pytest.xfail()
-        wfm, (x, y) = ymlSystem2PyWake(examples_data_path + 'wind_energy_system/IEA37_case_study_1_2_wind_energy_system.yaml',
+
+        plant_reference_path = Path(windIO_plant_lib.__file__).parent
+        wfm, (x, y) = ymlSystem2PyWake(plant_reference_path / 'wind_energy_system/IEA37_case_study_1_2_wind_energy_system.yaml',
                                        IEA37SimpleBastankhahGaussian)
         ref = np.array([9444.60012, 8497.90004, 11383.32869, 14173.40367,
                         20979.36776, 25590.86774, 39252.85757, 43197.65856,
@@ -35,7 +57,9 @@ except ModuleNotFoundError:
     def test_IEA37_case_study1_16wt_TopFarm2():
         if topfarm is None:
             pytest.xfail()
-        tf = ymlSystem2TopFarm(examples_data_path + 'wind_energy_system/IEA37_case_study_1_2_wind_energy_system.yaml')
+
+        plant_reference_path = Path(windIO_plant_lib.__file__).parent
+        tf = ymlSystem2TopFarm(plant_reference_path / 'wind_energy_system/IEA37_case_study_1_2_wind_energy_system.yaml')
         wfm = tf.cost_comp.windFarmModel
         wfm.verbose = False
         wfm.site.default_wd = np.arange(0, 360, 22.5)
